refactor: 쓰레기 퍼즐맞추기 미션 Mock api 수정 (#119)

+++ src/main/java/com/sesacthon/foreco/mock/mission/dto/QuizMissionChoice.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 @Getter
 public class QuizMissionChoice {
 
-  Integer id;
+  Long id;
   String name;
 
-  public QuizMissionChoice(Integer id, String name) {
+  public QuizMissionChoice(Long id, String name) {
     this.id = id;
     this.name = name;
   }
